changed double to float

+++ feature_extraction/ecg_features.h
@@ -5,11 +5,11 @@
 using namespace std;
 
 struct pvc_features {
-    double pre_rr, post_rr, r_amp;
-    double pr_interval, qrs_interval, qt_interval, qrs_area;
+    float pre_rr, post_rr, r_amp;
+    float pr_interval, qrs_interval, qt_interval, qrs_area;
 };
 
-vector<pvc_features> extract_beat_features(const vector<double>& r_peak_times,const vector<double>& r_amplitudes);
-int get_q_index (const vector<double>& filtered_ecg, int sampling_freq, double threshold_voltage, int r_index);
+vector<pvc_features> extract_beat_features(const vector<float>& r_peak_times,const vector<float>& r_amplitudes);
+int get_q_index (const vector<float>& filtered_ecg, int sampling_freq, float threshold_voltage, int r_index);
 
 #endif
