+++ signal_export/ecg_export.h
@@ -8,6 +8,6 @@
 using namespace std;
 
 
-void export_to_csv(const vector<vector<double>>& physical_ecg_data, int nsig, const string& filename);
+void export_to_csv(const vector<vector<float>>& physical_ecg_data, int nsig, const string& filename);
 
 #endif
