+++ signal_process/ecg_ht.h
@@ -6,9 +6,9 @@
 #include <algorithm>
 using namespace std;
 
-vector<double> ht_differentiation(const vector<double>& phys_ecg_data, int sampling_freq);
-vector<double> ht_squaring(const vector<double>& differentiated_ecg);
-vector<double> ht_moving_average(const vector<double>& squared_ecg, int window_size);
-vector<double> ht_adaptive_threshold(const vector<double>& averaged_ecg, double threshold_factor, int sampling_rate);
-vector<double> ht_peak_detection(const vector<double>& threshold_ecg, const vector<double>& filtered_ecg, int window_size);
+vector<float> ht_differentiation(const vector<float>& phys_ecg_data, int sampling_freq);
+vector<float> ht_squaring(const vector<float>& differentiated_ecg);
+vector<float> ht_moving_average(const vector<float>& squared_ecg, int window_size);
+vector<float> ht_adaptive_threshold(const vector<float>& averaged_ecg, float threshold_factor, int sampling_rate);
+vector<float> ht_peak_detection(const vector<float>& threshold_ecg, const vector<float>& filtered_ecg, int window_size);
 #endif
